Move MainCoroutineRule to core-common-test
