+++ app/src/test/kotlin/com/hedvig/app/feature/genericauth/GenericAuthViewModelTest.kt
@@ -4,9 +4,9 @@ import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isNotNull
 import assertk.assertions.isNull
+import com.hedvig.android.core.common.test.MainCoroutineRule
 import com.hedvig.android.market.Market
 import com.hedvig.android.market.MarketManager
-import com.hedvig.app.util.coroutines.MainCoroutineRule
 import com.hedvig.authlib.AuthAttemptResult
 import com.hedvig.authlib.AuthRepository
 import com.hedvig.authlib.AuthTokenResult
