+++ app/src/test/kotlin/com/hedvig/app/feature/loggedin/WhatsNewTest.kt
@@ -5,13 +5,13 @@ import assertk.assertThat
 import assertk.assertions.isNotNull
 import com.apollographql.apollo3.annotations.ApolloExperimental
 import com.apollographql.apollo3.mockserver.enqueue
+import com.hedvig.android.core.common.test.MainCoroutineRule
 import com.hedvig.app.apollo.runApolloTest
 import com.hedvig.app.apollo.toJsonStringWithData
 import com.hedvig.app.feature.whatsnew.WhatsNewRepository
 import com.hedvig.app.feature.whatsnew.WhatsNewViewModel
 import com.hedvig.app.feature.whatsnew.WhatsNewViewModelImpl
 import com.hedvig.app.testdata.feature.loggedin.WHATS_NEW
-import com.hedvig.app.util.coroutines.MainCoroutineRule
 import giraffe.WhatsNewQuery
 import giraffe.type.Locale
 import io.mockk.every
