+++ app/src/test/kotlin/com/hedvig/app/feature/marketing/ui/MarketingViewModelTest.kt
@@ -8,6 +8,7 @@ import assertk.assertions.isNotEqualTo
 import assertk.assertions.isNotNull
 import assertk.assertions.isTrue
 import assertk.assertions.prop
+import com.hedvig.android.core.common.test.MainCoroutineRule
 import com.hedvig.android.hanalytics.featureflags.FeatureManager
 import com.hedvig.android.market.Language
 import com.hedvig.android.market.Market
@@ -15,7 +16,6 @@ import com.hedvig.app.feature.marketing.MarketingViewModel
 import com.hedvig.app.feature.marketing.data.GetInitialMarketPickerValuesUseCase
 import com.hedvig.app.feature.marketing.data.GetMarketingBackgroundUseCase
 import com.hedvig.app.feature.marketing.data.UpdateApplicationLanguageUseCase
-import com.hedvig.app.util.coroutines.MainCoroutineRule
 import com.hedvig.hanalytics.HAnalytics
 import io.mockk.coEvery
 import io.mockk.mockk
